Copy over changes from dev_mode bootstrapping to examples/federated
After we shifted our focus from the federated example to the dev_mode, we continued to fix bugs and more subtle issues with the logic. This copies that logic back over to the example.

It also fixes a few things we noticed with the dev_mode index.js, such as it did not have the Promise.allSettled polyfill defined for older browsers.