Fill in background color above and below a span with an ansi background color.
This eliminates the gaps between lines where you have an ansi background color, which is caused by the lines having a line height that is slightly larger than the font size so that we have some nice spacing between the lines. When we have ANSI background colors, that spacing becomes really glaring, as noted in #7760.

This is yet another solution to #7760 that avoids the problems we see from the solution in #8555 (noted in #9600).

Fixes #9600