Enable large file uploads for Jupyter Server
In #9616, I did not realize that JupyterLab still supports the Jupyter Notebook server. This fix supports Jupyter Notebook as a server.

Jupyter Server advertises itself as notebook version 1.2 or so at the time of this commit. JupyterLab 2 supports Jupyter Notebook >=4.3.1. Thus we assume that if we get a version number less than 4, we are actually talking to Jupyter Server, and chunked uploading is supported.

Eventually this code should be removed when JupyterLab no longer supports Jupyter Notebook server, and hopefully that is before Jupyter Server 4.0.

Fixes #9626