Use Path.resolve() to get canonical case-sensitive path names.
Fixes #9698

This fixes `jupyter lab build` on Windows when using pip. The underlying problem is in pip (from its dependency distlib) and how it normalizes paths to lowercase when it creates executables for Windows entry points (see https://bitbucket.org/pypa/distlib/issues/143/do-not-normcase-executable-path-when and https://github.com/pypa/pip/issues/6582#issuecomment-770197386). Basically, when we set the current working directory to a lower-case version of the path, node starts using the lower-case version for its module cache when building JupyterLab. However, the mini-css-extract plugin in JLab asks node to resolve a path for its loader, which uses the canonical case for the path. This leads to Node having two copies of the same modules in its cache (differentiated by the path case), which leads to problems when doing instanceof checks. See the exploration at https://github.com/jupyterlab/jupyterlab/issues/9698 for more details.

Until the issue is fixed in distlib/pip, we can work around it by using Path.resolve(), which seems to give back a canonical name (including case) for a path. It’s possible that we should use this many more places in the code, but it seems that this change is minimal enough to at least get the windows build working, and I want to be conservative here.

We initially tried using os.path.realpath, and while that works on python 3.9, it does not correct case on python 3.6.

You can see the effect of this change by installing jupyterlab with (current) pip on Windows and running `jupyter lab path`. You’ll notice that the application directory is lowercase before this commit, but normal case after (e.g., the drive letter is uppercase, etc.).