Update extension_dev.rst
I just stumbled over that sentence while reading up on JupyterLab extensions. At first it sounds as if source extensions and prebuilt extensions were mutually exclusive types of extensions, but two paragraphs later it is clarified that the same source base of an extension can be used to publish of both "types" of extension at the same time. As the audience of this documentation is developers, it would be better to take the developer perspective, not the end-user perspective.