Updates the locking configuration
Our previous locking solution is now in maintenance mode (https://github.com/dessant/lock-threads-app). The successor is a workflow. This converts our locking configuration to use the workflow version.

I also increased the inactive timer to 180 days. Sometimes I’ve noticed we want to refer to a recent PR, but the backlink is not put in because the PR is locked. This lets us easily refer to a PR up to 6 months old and have the auto-backlinking work well.