Remove the auto-switch to “mobile” mode, and behavior switches associated with mobile mode
Fixes #9567

Reverts part of #8456

See https://github.com/jupyterlab/jupyterlab/issues/9567 for an extensive discussion about this. In short, the existing auto-switching behavior was disruptive to desktop users, and conflated issues of responsive layout, mobile platform, and simple interface. The plan is to take a step back and design an approach that disentangles each of these concerns better.