Do not display the splash screen in the application state plugin
The splash screen is displayed in the theme manager, so displaying it in the state plugin means the splash screen is displayed twice (visually this often appears as the Jupyter moons orbiting, then stopping and starting over again). The state manager does not know if the theme is light or dark, so does not know the right options for displaying the splash screen anyway.