Change extension manager to only request from npm when warning has been disclaimed.
Before this commit, the extension manager was always requesting a search from npm, even if the warning had not been disclaimed. This meant that a default JLab install (with extension manager enabled) would always make an external network request.

This changes things so that the network request is only made if the warning has been disclaimed. It actually does more than that, and also cuts out the request to the server, but since the installed packages are also not shown when the disclaimer has not been agreed to, it is okay to prevent that request as well.