Add a clarifying comment for the download plugin.
This expands on the changes in #10066, where the download entries are separated into a separate plugin that can be disabled. There have been a fair number of requests to disable downloading, which is impossible to do without server changes. This commit adds a comment clarifying that disabling this new plugin is more of a UX change rather than a functional change.

For more conversation around this, see #5274.