fix: remove the 3-second startup delay of the kernel connection (#10321)
The logic for immediately sending a kernel info message on restart
is also applied when starting. The kernel info request in the
constructor is removed because it is redundant (this message is
also send on the web socket open event) and causes a
"status is not connected" console error on startup in the new
situation.