Update reference UI screenshots to reflect a single launcher tab not having a close “X”.
In the latest 3.0 release, if a single launcher tab is open, it does not have a close X in the tab (since we must have at least one tab open). In master, the close X does appear, but does nothing when clicking on it. After  #10516, the 3.0 behavior is restored: if a single launcher tab is open, it does not have an X.