Update ui test screenshots that were slightly shifted.
As seen in the previous commit, this PR changed a menu item from “Auto Close Brackets for Text Editor” to “Auto Close Brackets”, which made the menu slightly narrower, which in turn slightly shifted the submenus. This resulted in the screen capture being slightly shifted compared to the reference screenshots.