Add command to display relevant key bindings
Determine relevant keybindings for the current active element (taking into account the convention that matching stops when an element matches [data-lm-suppress-shortcuts] or [data-p-suppress-shortcuts]) and display them in the console. This is a step towards showing the user all relevant keybindings for wherever they are.

We still need to display these bindings in a way the user will see.

We also see that some existing core keybindings do not have labels, and others do not have captions. This needs to be fixed as well.

In order to activate this command, add this user keyboard shortcut:

{command: "apputils:display-shortcuts",
 keys: ["Ctrl Shift H"],
 selector: "body"}

See https://github.com/jupyterlab/jupyterlab/issues/10284