Make keyboard shortcut discovery logic more complete
If two keyboard shortcuts have the same keys, use the Lumino rules to
break ties, namely that the shortcut that is closer to active element
wins, and if they are the same distance, the one with greater specificity
wins, and if they are the same specificity, the later one wins (i.e.,
the one added most recently).

Additionally, display the shortcuts grouped by distance from the active
element. This approximately groups shortcuts into most specific to least
specific.

We may want to eventually move some or all of this logic into apputils
and just call it from apputils-extension.