Building extensions fail if not using latest patch
Due to my changes in https://github.com/jupyterlab/jupyterlab/pull/12533
which were included in v3.4.1 extension builds may fail without an explicit
upgrade to the new patch (should not be necessary for patches).

This is because my change will compare against the exact version installed
rather than the range defined in the package.json, this update makes the
new check conditional to the original failure it was addressing.