Make password inputs not give away how many characters were typed
This aligns with classic Jupyter Notebook behavior, and is generally a good security idea.

Also, this decouples the implementation of the HTML input widget and the notion of the input being a password input by storing the password field as a private attribute of the stdin control, rather than relying on how the input was created.