Optimize `eachLine` loops (#12746)
When modifying each line of a large document it is very important that we do not update the DOM after each update. In CodeMirror this is avoided by wrapping things in operations.

For `clearGutter` we just call the already defined method in CM, which uses an operation by default. It might also be slightly more optimized in other ways.