Debugger: Make kernel source list react based (#12751)
* Make kernel sources filter sticky

This way it will be visible even if you scroll a lot!

* Make kernel source list react based

This improves performance when updating of the kernel sources, which happens whenever a document that is being debugged gets focused. This includes when switching from the notebook document to a kernel source editor. Ideally the kernel sources would not be refreshed when switching between tabs that share the same debug context.