Adding default parameter to reload method in Galata
Galata has overrode the page.reload method of Playwright and as a result has introduced assertions that are only true if the JupyterLab Launcher is visible after reload.

The default parameter is a boolean named `waitForLauncher` which is set to true by default. If you wish to bypass the assertions that the Launcher is visible, you can set it equal to false.

This was by the suggestion of @fcollonval on the issue: https://github.com/jupyterlab/jupyterlab/issues/12924