Update Python icon to be PSF Trademark complient
The current python icon was custom made for JupyterLab in it's early
alpha phase and doesn't comply with the PSF Trademark.

This replaces that icon with one derived from the offical icon svg as
provided at https://www.python.org/community/logos/

In order to support use as a LabIcon and use JupyterLab styling the
svg was drastically simlified. The remaining svg uses the same paths
and the same colors, but without the opacity gradient.

Fixes #13023