Optimize text mimerenderer: ansi vs autolink
This optimizes the rendering for cases where there are many more elements of one type than the other. E.g. for a large syntax highlighted source file. In this case there can be many (tens of) thousands of span elements, but likely only a few links. In this case, multiple span elements will all correspond to a single text node in the autolink output. In these cases, we can simply use the nodes from the span element list, without splitting / cloning any nodes. This greatly reduces the time spent.