Link the background colors and the ui font colors in the relevant css.
Update menu.css.

Update menu.css to revert the change on the background color of lm-MenuBar-item.lm-mod-disabled.

Set the font color for jp-Launcher-sectionTitle to jp-ui-font-color0 to match the background color index.

Propose updates in the help extension, in the toolbar of the notebook, in the setting editor and in the dialog box.

Revert changes that were decreasing contrast in the toolbar, in the statusbar and in the tabs bar.*

Revert change on switch label color.

Update packages/rendermime/style/base.css

Co-authored-by: Frédéric Collonval <fcollonval@users.noreply.github.com>

Update packages/settingeditor/style/base.css

Co-authored-by: Frédéric Collonval <fcollonval@users.noreply.github.com>