Make changes to the tooltips of the toolbar actions.
Add a selectedCells attribute in Notebook to return an array of active or selected cells and be able to adapt the caption to singular or plural version depending on the number of cells that are duplicated.

Add a notifyCommandChanged on selectionChanged to fix the issue on the update of the list of active or selected cells.

Update packages/notebook-extension/src/index.ts, cache the selectedCells list, apply the same logics for singular and plural in the labels and captions to all the commands involving selected cells and add the number of selected cells in the caption for the relevant commands.

Update packages/notebook-extension/src/index.ts and packages/notebook/src/widget.ts.

Apply suggestions from code review

Co-authored-by: Florence Haudin <haudin.florence@gmail.com>