Update README.md
Updated README.md file to help new users better understand how to update the path, what the path is, and what it means for them moving forward. This commit was created in the hopes of lowering the barrier to entry so users who are not as computer savvy, but are more research inclined, may have an easier transition into jupyter notebooks.