Solve #1947 in a more specific way
The keydown->ensureFocus code (#1958) was added to fix #1947.

The idea was that any key event that bubbles up from within the
notebook widget but causes the focus to move somewhere outside
the notebook was something like when the user presses enter from
within a Stdin widget, causing this widget, which had focus, to
remove itself from the DOM, causing focus to get lost, meaning
returned to the body node.

But if we change the _ensureFocus method so that it ensures
focus on the active cell rather than on the notebook as a whole
it can lead to bugs like the one where the user presses tab
while on the notebook footer, triggering the keydown->ensureFocus
code, causing focus to go up the notebook, backwards (because
all cells are above the footer), which makes it impossible to tab
forward out of the notebook. Because you get into the following
loop:

cell -> ... -> footer (-> keydown->ensureFocus) -> cell -> ... -> footer ->