Fix `_updateConstraints` missing `protected`/`private` classifier.
It was meant to be private - simple typo in a previous PR.
To avoid breakages in case if anyone has adopted it in a subclass
I am marking it as `protected`. If anyone was using it as a public
method this was an error - it should not be used like that.