Keep active cell rendered in the `full` windowed mode (#15286)
* Initial work on ensuring the active cell is always rendered

* Scroll to the active cell on `keydown` if out of view

* Document and move implementation to private methods,

also collapse padding to ensure bounding box is 0.

* Add browser tests, update existing tests

* Extend the scroll test to markdown cells

* Preserve active cell for all types of cells

* Fix focus for soft-hidden code cells

* Move logic around to actually detach non-active markdown cells

* Do not re-attach if cell was soft-hidden

So this was caught by tests: if a markdown cell a heading
the logic in `showEditor()` would move cursor to the heading
start; while that logic maybe should be improved, in this
case the problem was that we went ahead with attaching already
attached cell; this commit corrects it.

* Simplify the scroll test to allow soft-hidden cells

The cells which are soft hidden will be interpreted
as in `hidden` state by playwright, but won't have
a falsy bounding box. The second set of assertions
using falsy/truthy bbox is thus not good for soft-
hidden cells, and since it is redundant it gets dropped.