Simplify the logic for resetting all shortcuts
Previously it would loop through top-level properties
and call `setting.remove()` on them. There is only one
top-level property in shortcuts settings, a list called
`settings`. The `removeShortCut` method was thus
a misnomer (it was removing a settings key instead).
In the future if other keys were introduced, e.g.
allowing to configure shortcuts UI, the old logic
would wrongly reset those as well.

Also, this makes testing easier as there are
fewer things to provide in the external bundle.