Do not capture the full main area in galata notebook tests
it does not make much sense, especially for cells; instead,
capture the notebook itself. Add a new test for saving,
because that was implicitly tested by `Run Cells` test
(and is not anymore). Of note, one of the tests,
`Open and run cell by cell` is still taking a snapshot
of the entire panel; this is intentional - we do want
one snapshot to be more encompassing (but not all of them!).