Add Signal to proxy all Completer selected Signals
Completer currently emits a signal when a completion is selected.
This surfaces that signal to the CompletionProviderManager, which
then proxies all the Completer signals to the new signal.

This will allow extensions to know when a selection is made in a
Completer. Currently there are edge cases where a selection can be
made without calling a command that can be connected to.

Signed-off-by: Alex Bozarth <ajbozart@us.ibm.com>