Do not skip cells with hidden input (fixed #11162)
This fixes the issue where cells with hidden (i.e. "folded") code are not accessible
from the cell navigation keyboard shortcuts "k" "j" "<Up>" and "Down". This makes
running or navigating to folded cells impossible without either 1) clicking the cell
(which also automatically unfolds the cell) or 2) using the shift + enter shortcut
(which requires running preceding cells). Sometimes the user may already know the
contents of a hidden cell by its output or preceding markdown headers but may wish
to keep the cell folded for the sake of readability or to prevent large multi-line
cells from dominating the screen.