hub-extension: Restart server with iframe of spawn page
Instead of opening the hub spawn page in a new tab, which then will
redirect to open a new workspace, create an iframe for the spawn
page and then close it when it redirects to the lab url.  This will
maintain most of the state during restarts, although kernels will
of course be restarted and any open terminals will be closed.

Right now, this uses a naive check for the redirection to the lab
URL.  We should build a better check based on path.urls.