hub-extension: Fall back on iframe error
If the iframe enclosure of the spawn page raises an error for some
reason, open the spawn page in a new window, as before.  This will
catch the problem when the spawn page can't be loaded due to headers
blocking embedding, as a DOMException is raised when we try to read
iframe location.href.