Fix `ObservableList.removeValue` behaviour: if item isn't found, don't just remove the last item! (#18019)
* If item isn't found, don't just remove the last item!

ArrayExt.findFirstIndex will return -1 if the value is not in the array. Without this check, this.remove with index=-1 will splice off the last item from the array, which is undesirable since we weren't trying to remove this item (it doesn't match, else the index would have given this.length - 1).

* Added test for fixed removeValue behavior

A test verifying the final element isn't removed when removeValue is called with a nonexistent item.