Use the browser-native fetch api to decode base64 urls into blobs.
This also opens up the possibility of decoding blob urls.

One restriction we have now is that we only allow image/audio/video uploads through data or blob urls. We figured this was safe to try at the start, and we can relax this as desired. We also may open up to arbitrary urls rather than just local data/blob urls.